refactor(vsock-test): split integration tests by protocol area
