+++ crates/vsock-test/tests/integration/main.rs
@@ -1,0 +1,14 @@
+#![allow(
+    clippy::unwrap_used,
+    clippy::expect_used,
+    clippy::panic,
+    clippy::indexing_slicing,
+    clippy::unreachable
+)]
+
+mod exec;
+mod shutdown;
+mod spawn_watch;
+mod stdout_streaming;
+mod support;
+mod write_file;
